Final tests fixup
